feat: move and rename request dirs/files
