+++ web-client/src/modules/requests/components/NewRequest.tsx
@@ -5,7 +5,7 @@ import { StepForwardButton } from 'src/components/Buttons';
 import TitleWithOrangeUnderline from 'src/components/TitleWithOrangeUnderline/TitleWithOrangeUnderline';
 import styled from 'styled-components';
 
-import SearchIcon from '../../assets/search.svg';
+import SearchIcon from '../assets/search.svg';
 import { RequestInput } from './RequestReview';
 
 const { Option } = Select;
