+++ web-client/src/modules/requests/constants.tsx
@@ -1,0 +1,5 @@
+import Location from 'react-app-location';
+
+export const FindRequestsLocation = new Location('/requests/find');
+export const NewRequestsLocation = new Location('/requests/new');
+export const TabbedRequestsLocation = new Location('/requests/list');
