+++ web-client/src/modules/requests/containers/OfferedRequestsContainer.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from 'react';
 
-export const OffersContainer: React.FC = (): ReactElement => (
+export const OfferedRequestsContainer: React.FC = (): ReactElement => (
   <span>Offered Requests</span>
 );
