--- web-client/src/modules/requests/pages/routes/ArchivedRequestsRoute/ArchivedRequestsRoute.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import ClosedRequestsContainer from '../../../containers/ArchivedRequestsContainer/ArchivedRequestsContainer';
-
-const ClosedRequestsRoute: React.FC = () => <ClosedRequestsContainer />;
-
-export default ClosedRequestsRoute;
