--- web-client/src/modules/requests/pages/routes/ArchivedRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const ArchivedRequestsLocation = new Location('/requests/archived');
