--- web-client/src/modules/requests/pages/routes/FindRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const FindRequestsLocation = new Location('/requests/find');
