--- web-client/src/modules/requests/pages/routes/FinishedRequestsRoute/FinishedRequestsRoute.tsx
@@ -1,8 +1,0 @@
-import React, { ReactElement } from 'react';
-
-import FinishedRequestsContainer from '../../../containers/FinishedRequestsContainer/FinishedRequestsContainer';
-
-const CompletedRequestsRoute: React.FC = (): ReactElement => (
-  <FinishedRequestsContainer />
-);
-export default CompletedRequestsRoute;
