--- web-client/src/modules/requests/pages/routes/FinishedRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const FinishedRequestsLocation = new Location('/requests/finished');
