--- web-client/src/modules/requests/pages/routes/OngoingRequestsRoute/OngoingRequestsRoute.tsx
@@ -1,8 +1,0 @@
-import React, { ReactElement } from 'react';
-
-import OngoingRequestsContainer from '../../../containers/OngoingRequestsContainer/OngoingRequestsContainer';
-
-const OngoingRequestsRoute: React.FC = (): ReactElement => (
-  <OngoingRequestsContainer />
-);
-export default OngoingRequestsRoute;
