--- web-client/src/modules/requests/pages/routes/OngoingRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const OngoingRequestsLocation = new Location('/requests/ongoing');
