--- web-client/src/modules/requests/pages/routes/OpenRequestsRoute/OpenRequestsRoute.tsx
@@ -1,8 +1,0 @@
-import React, { ReactElement } from 'react';
-
-import OpenRequestsContainer from '../../../containers/OpenRequestsContainer/OpenRequestsContainer';
-
-const OpenRequestsRoute: React.FC = (): ReactElement => (
-  <OpenRequestsContainer />
-);
-export default OpenRequestsRoute;
