--- web-client/src/modules/requests/pages/routes/OpenRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const OpenRequestsLocation = new Location('/requests/open');
